Conflicts:
	dev/Basic/shared/conf/simpleconf.cpp
	dev/Basic/shared/entities/misc/TripChain.hpp
	dev/Basic/shared/geospatial/Lane.hpp
	dev/Basic/shared/geospatial/MultiNode.hpp
	dev/Basic/shared/geospatial/Node.hpp
	dev/Basic/shared/geospatial/RoadSegment.cpp
	dev/Basic/shared/geospatial/xmlLoader/geo8-driver.cpp
	dev/Basic/shared/geospatial/xmlLoader/geo8-driver.hpp
	dev/Basic/shared/geospatial/xmlLoader/geo8-pimpl.cpp

+++ dev/visualizer/Large Network Visualizer/src/sim_mob/vis/network/BusStop.java
@@ -3,14 +3,9 @@ package sim_mob.vis.network;
 
 import java.awt.*;
 import java.awt.geom.Rectangle2D;
-import java.awt.image.BufferedImage;
 
-
-import sim_mob.vect.SimpleVectorImage;
-import sim_mob.vis.MainFrame;
 import sim_mob.vis.controls.DrawParams;
 import sim_mob.vis.controls.DrawableItem;
-import sim_mob.vis.network.basic.FlippedScaledPoint;
 import sim_mob.vis.network.basic.ScaledPoint;
 import sim_mob.vis.util.Utility;
 
@@ -61,8 +56,8 @@ public class BusStop implements DrawableItem {
 	private static Stroke BusStopStroke = new BasicStroke(2.0F);
 		
 	private boolean isUni;   //Rather than having multiple classes....
-	private Integer id;
-	public BusStop (ScaledPoint nearOne, ScaledPoint nearTwo, ScaledPoint farOne, ScaledPoint farTwo, Integer id) {
+	private Long id;
+	public BusStop (ScaledPoint nearOne, ScaledPoint nearTwo, ScaledPoint farOne, ScaledPoint farTwo, Long id) {
 
 	
 		this.id = id;
@@ -95,7 +90,7 @@ public class BusStop implements DrawableItem {
 	public boolean getIsUni() {
 		return isUni;
 	}
-	public Integer getID(){
+	public Long getID(){
 		return id;
 	}
